+++ src/app/providers/index.ts
@@ -1,0 +1,4 @@
+import { router } from "./router/router";
+import { store } from "./store/store";
+
+export { router, store };
