+++ src/features/Favorites/actions/FavoritesActions.ts
@@ -1,6 +1,6 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { arrayRemove, arrayUnion, doc, updateDoc } from "firebase/firestore";
-import { db } from "../../../firebase";
+import { db } from "../../../app/firebase/firebase";
 
 interface UserFavorites {
   email: string;
