+++ src/features/Search/SearchSelectors/SearchSelector.ts
@@ -1,4 +1,4 @@
-import { RootState } from "../../../app/providers/store";
+import { RootState } from "../../../app/providers/store/store";
 
 export const getSearchMovies = (state: RootState) => state.search.movies;
