+++ src/main.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
-import { store } from "./app/providers/store/index.tsx";
 import { RouterProvider } from "react-router-dom";
-import router from "./app/providers/router/router.tsx";
+import { router, store } from "./app/providers";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
