--- src/shared/UI/Button/Button.tsx
@@ -1,9 +1,0 @@
-import { Props } from "./Button.type";
-
-export function Button({ children, onClick, type }: Props) {
-  return (
-    <button type={type} onClick={onClick}>
-      {children}
-    </button>
-  );
-}
