--- src/shared/UI/Button/Button.type.ts
@@ -1,7 +1,0 @@
-import { ReactNode } from "react";
-
-export type Props = {
-  children?: ReactNode;
-  type: "button" | "submit";
-  onClick?: React.MouseEventHandler<HTMLButtonElement>;
-};
